+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `ec2_ami` `state` parameter validation correction
 * `get_url`, `unarchive` support for `owner`, `group`, `mode`, `setype`, `selevel`
 * `wait_for` module `port` parameters validation fix - integer
+* `firewalld` permanent type fix
 
 # 1.0.6 - 14 October 2016
 * Allow sharing variables between tasks (#29)